refactor(util): create shared utils
