+++ util/games.mjs
@@ -7,148 +7,146 @@
  * @property {string} [abbr]
  */
 
-export default {
-    /** @type {Object.<string, Game>} */
-    games: {
-        cs_CZ: {
-            flag: '🇨🇿',
-            name: 'Operacni-Stredisko.cz',
-            shortURL: 'operacni-stredisko.cz',
-            police: 'policie',
-        },
-        da_DK: {
-            flag: '🇩🇰',
-            name: 'Alarmcentral-Spil.dk',
-            shortURL: 'alarmcentral-spil.dk',
-            police: 'politi',
-        },
-        de_DE: {
-            flag: '🇩🇪',
-            name: 'Leitstellenspiel.de',
-            shortURL: 'leitstellenspiel.de',
-            police: 'polizei',
-            abbr: 'LSS',
-        },
-        en_AU: {
-            flag: '🇦🇺',
-            name: 'Missionchief-Australia.com',
-            shortURL: 'missionchief-australia.com',
-            police: 'police',
-            abbr: 'MC',
-        },
-        en_GB: {
-            flag: '🇬🇧',
-            name: 'Missionchief.co.uk',
-            shortURL: 'missionchief.co.uk',
-            police: 'police',
-            abbr: 'MC',
-        },
-        en_US: {
-            flag: '🇺🇸',
-            name: 'Missionchief.com',
-            shortURL: 'missionchief.com',
-            police: 'police',
-            abbr: 'MC',
-        },
-        es_ES: {
-            flag: '🇪🇸',
-            name: 'Centro-de-Mando.es',
-            shortURL: 'centro-de-mando.es',
-        },
-        es_MX: {
-            flag: '🇲🇽',
-            name: 'Centro-de-Mando.mx',
-            shortURL: 'centro-de-mando.mx',
-        },
-        fi_FI: {
-            flag: '🇫🇮',
-            name: 'Hätäkeskuspeli.com',
-            shortURL: 'hatakeskuspeli.com',
-            police: 'poliisi',
-        },
-        fr_FR: {
-            flag: '🇫🇷',
-            name: 'Operateur112.fr',
-            shortURL: 'operateur112.fr',
-            police: 'police',
-        },
-        it_IT: {
-            flag: '🇮🇹',
-            name: 'Operatore112.it',
-            shortURL: 'operatore112.it',
-            police: 'polizia',
-        },
-        ja_JP: {
-            flag: '🇯🇵',
-            name: 'Missionchief-Japan.com',
-            shortURL: 'missionchief-japan.com',
-        },
-        ko_KR: {
-            flag: '🇰🇷',
-            name: 'Missionchief-Korea.com',
-            shortURL: 'missionchief-korea.com',
-        },
-        nb_NO: {
-            flag: '🇳🇴',
-            name: 'Nodsentralspillet.com',
-            shortURL: 'nodsentralspillet.com',
-            police: 'politiet',
-        },
-        nl_NL: {
-            flag: '🇳🇱',
-            name: 'Meldkamerspel.com',
-            shortURL: 'meldkamerspel.com',
-            police: 'politie',
-        },
-        pl_PL: {
-            flag: '🇵🇱',
-            name: 'Operatorraunkowy.pl',
-            shortURL: 'operatorratunkowy.pl',
-            police: 'policja',
-        },
-        pt_BR: {
-            flag: '🇧🇷',
-            name: 'Operador193.com',
-            shortURL: 'operador193.com',
-        },
-        pt_PT: {
-            flag: '🇵🇹',
-            locale_fallback: 'en_US',
-            name: 'Jogo-operador112.com',
-            shortURL: 'jogo-operador112.com',
-            police: 'policia',
-        },
-        ro_RO: {
-            flag: '🇷🇴',
-            name: 'Jocdispecerat112.com',
-            shortURL: 'jocdispecerat112.com',
-        },
-        ru_RU: {
-            flag: '🇷🇺',
-            name: 'Dispetcher112.ru',
-            shortURL: 'dispetcher112.ru',
-        },
-        sk_SK: {
-            flag: '🇸🇰',
-            locale_fallback: 'en_US',
-            name: 'Dispečerskom-centrum.com',
-            shortURL: 'dispecerske-centrum.com',
-        },
-        sv_SE: {
-            flag: '🇸🇪',
-            name: 'Larmcentralen-spelet.se',
-            shortURL: 'larmcentralen-spelet.se',
-            police: 'polis',
-        },
-        tr_TR: {
-            flag: '🇹🇷',
-            name: '112-Merkez.com',
-            shortURL: '112-merkez.com',
-        },
-        uk_UA: {
-            flag: '🇺🇦',
-            name: 'Dyspetcher101-game.com',
-            shortURL: 'dyspetcher101-game.com',
-        },
+/** @type {Object.<string, Game>} */
+export const GAMES = {
+    cs_CZ: {
+        flag: '🇨🇿',
+        name: 'Operacni-Stredisko.cz',
+        shortURL: 'operacni-stredisko.cz',
+        police: 'policie',
+    },
+    da_DK: {
+        flag: '🇩🇰',
+        name: 'Alarmcentral-Spil.dk',
+        shortURL: 'alarmcentral-spil.dk',
+        police: 'politi',
+    },
+    de_DE: {
+        flag: '🇩🇪',
+        name: 'Leitstellenspiel.de',
+        shortURL: 'leitstellenspiel.de',
+        police: 'polizei',
+        abbr: 'LSS',
+    },
+    en_AU: {
+        flag: '🇦🇺',
+        name: 'Missionchief-Australia.com',
+        shortURL: 'missionchief-australia.com',
+        police: 'police',
+        abbr: 'MC',
+    },
+    en_GB: {
+        flag: '🇬🇧',
+        name: 'Missionchief.co.uk',
+        shortURL: 'missionchief.co.uk',
+        police: 'police',
+        abbr: 'MC',
+    },
+    en_US: {
+        flag: '🇺🇸',
+        name: 'Missionchief.com',
+        shortURL: 'missionchief.com',
+        police: 'police',
+        abbr: 'MC',
+    },
+    es_ES: {
+        flag: '🇪🇸',
+        name: 'Centro-de-Mando.es',
+        shortURL: 'centro-de-mando.es',
+    },
+    es_MX: {
+        flag: '🇲🇽',
+        name: 'Centro-de-Mando.mx',
+        shortURL: 'centro-de-mando.mx',
+    },
+    fi_FI: {
+        flag: '🇫🇮',
+        name: 'Hätäkeskuspeli.com',
+        shortURL: 'hatakeskuspeli.com',
+        police: 'poliisi',
+    },
+    fr_FR: {
+        flag: '🇫🇷',
+        name: 'Operateur112.fr',
+        shortURL: 'operateur112.fr',
+        police: 'police',
+    },
+    it_IT: {
+        flag: '🇮🇹',
+        name: 'Operatore112.it',
+        shortURL: 'operatore112.it',
+        police: 'polizia',
+    },
+    ja_JP: {
+        flag: '🇯🇵',
+        name: 'Missionchief-Japan.com',
+        shortURL: 'missionchief-japan.com',
+    },
+    ko_KR: {
+        flag: '🇰🇷',
+        name: 'Missionchief-Korea.com',
+        shortURL: 'missionchief-korea.com',
+    },
+    nb_NO: {
+        flag: '🇳🇴',
+        name: 'Nodsentralspillet.com',
+        shortURL: 'nodsentralspillet.com',
+        police: 'politiet',
+    },
+    nl_NL: {
+        flag: '🇳🇱',
+        name: 'Meldkamerspel.com',
+        shortURL: 'meldkamerspel.com',
+        police: 'politie',
+    },
+    pl_PL: {
+        flag: '🇵🇱',
+        name: 'Operatorraunkowy.pl',
+        shortURL: 'operatorratunkowy.pl',
+        police: 'policja',
+    },
+    pt_BR: {
+        flag: '🇧🇷',
+        name: 'Operador193.com',
+        shortURL: 'operador193.com',
+    },
+    pt_PT: {
+        flag: '🇵🇹',
+        locale_fallback: 'en_US',
+        name: 'Jogo-operador112.com',
+        shortURL: 'jogo-operador112.com',
+        police: 'policia',
+    },
+    ro_RO: {
+        flag: '🇷🇴',
+        name: 'Jocdispecerat112.com',
+        shortURL: 'jocdispecerat112.com',
+    },
+    ru_RU: {
+        flag: '🇷🇺',
+        name: 'Dispetcher112.ru',
+        shortURL: 'dispetcher112.ru',
+    },
+    sk_SK: {
+        flag: '🇸🇰',
+        locale_fallback: 'en_US',
+        name: 'Dispečerskom-centrum.com',
+        shortURL: 'dispecerske-centrum.com',
+    },
+    sv_SE: {
+        flag: '🇸🇪',
+        name: 'Larmcentralen-spelet.se',
+        shortURL: 'larmcentralen-spelet.se',
+        police: 'polis',
+    },
+    tr_TR: {
+        flag: '🇹🇷',
+        name: '112-Merkez.com',
+        shortURL: '112-merkez.com',
+    },
+    uk_UA: {
+        flag: '🇺🇦',
+        name: 'Dyspetcher101-game.com',
+        shortURL: 'dyspetcher101-game.com',
     },
 };
